requirements.txt
Django==4.2.13
environs==11.0.0
django-admin-sortable2==2.2.1
django-tinymce==4.1.0
pillow==10.3.0
tqdm==4.66.4
requests==2.32.3
pydantic==2.7.4
django-debug-toolbar==4.4.5